requirements.txt
Django
django-cors-headers
djangorestframework
firebase-admin
djangorestframework-simplejwt
uvicorn==0.34.0
asgiref==3.8.1